Use absolute imports and tidy up
