+++ spacy/tests/parser/test_nn_beam.py
@@ -1,19 +1,23 @@
+# coding: utf8
 from __future__ import unicode_literals
+
 import pytest
 import numpy
 from thinc.api import layerize
-
-from ...vocab import Vocab
-from ...syntax.arc_eager import ArcEager
-from ...tokens import Doc
-from ...gold import GoldParse
-from ...syntax._beam_utils import ParserBeam, update_beam
-from ...syntax.stateclass import StateClass
+from spacy.vocab import Vocab
+from spacy.language import Language
+from spacy.pipeline import DependencyParser
+from spacy.syntax.arc_eager import ArcEager
+from spacy.tokens import Doc
+from spacy.syntax._beam_utils import ParserBeam, update_beam
+from spacy.syntax.stateclass import StateClass
+from spacy.gold import GoldParse
 
 
 @pytest.fixture
 def vocab():
     return Vocab()
+
 
 @pytest.fixture
 def moves(vocab):
@@ -85,3 +89,13 @@ def test_beam_advance(beam, scores):
 def test_beam_advance_too_few_scores(beam, scores):
     with pytest.raises(IndexError):
         beam.advance(scores[:-1])
+
+
+def test_beam_parse():
+    nlp = Language()
+    nlp.add_pipe(DependencyParser(nlp.vocab), name='parser')
+    nlp.parser.add_label('nsubj')
+    nlp.parser.begin_training([], token_vector_width=8, hidden_width=8)
+
+    doc = nlp.make_doc(u'Australia is a country')
+    nlp.parser(doc, beam_width=2)
