+++ spacy/tests/pipeline/test_textcat.py
@@ -1,7 +1,13 @@
 # coding: utf8
-
 from __future__ import unicode_literals
-from ...language import Language
+
+import random
+import numpy.random
+from spacy.language import Language
+from spacy.pipeline import TextCategorizer
+from spacy.vocab import Vocab
+from spacy.tokens import Doc
+from spacy.gold import GoldParse
 
 
 def test_simple_train():
@@ -16,3 +22,35 @@ def test_simple_train():
     doc = nlp(u'aaa')
     assert 'answer' in doc.cats
     assert doc.cats['answer'] >= 0.5
+
+
+def test_textcat_learns_multilabel(en_vocab):
+    random.seed(5)
+    numpy.random.seed(5)
+    docs = []
+    letters = ['a', 'b', 'c']
+    for w1 in letters:
+        for w2 in letters:
+            cats = {letter: float(w2==letter) for letter in letters}
+            docs.append((Doc(en_vocab, words=['d']*3 + [w1, w2] + ['d']*3), cats))
+    random.shuffle(docs)
+    model = TextCategorizer(en_vocab, width=8)
+    for letter in letters:
+        model.add_label(letter)
+    optimizer = model.begin_training()
+    for i in range(30):
+        losses = {}
+        Ys = [GoldParse(doc, cats=cats) for doc, cats in docs]
+        Xs = [doc for doc, cats in docs]
+        model.update(Xs, Ys, sgd=optimizer, losses=losses)
+        random.shuffle(docs)
+    for w1 in letters:
+        for w2 in letters:
+            doc = Doc(en_vocab, words=['d']*3 + [w1, w2] + ['d']*3)
+            truth = {letter: w2==letter for letter in letters}
+            model(doc)
+            for cat, score in doc.cats.items():
+                if not truth[cat]:
+                    assert score < 0.5
+                else:
+                    assert score > 0.5
