+++ spacy/tests/tokenizer/test_whitespace.py
@@ -1,7 +1,4 @@
 # coding: utf-8
-"""Test that tokens are created correctly for whitespace."""
-
-
 from __future__ import unicode_literals
 
 import pytest
