+++ spacy/tests/vocab_vectors/test_lexeme.py
@@ -1,9 +1,9 @@
 # coding: utf-8
 from __future__ import unicode_literals
 
-from ...attrs import *
-
 import pytest
+from spacy.attrs import IS_ALPHA, IS_DIGIT,
+
 
 @pytest.mark.parametrize('text1,prob1,text2,prob2', [("NOUN", -1, "opera", -2)])
 def test_vocab_lexeme_lt(en_vocab, text1, text2, prob1, prob2):
@@ -69,4 +69,3 @@ def test_lexeme_bytes_roundtrip(en_vocab):
     assert one.orth == alpha.orth
     assert one.lower == alpha.lower
     assert one.lower_ == alpha.lower_
-
